+++ lib/DependencyInjection/VSApiExtension.php
@@ -4,13 +4,16 @@ use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
+use Symfony\Component\DependencyInjection\Extension\PrependExtensionInterface;
 
 /** @experimental */
-final class VSApiExtension extends Extension
+final class VSApiExtension extends Extension implements PrependExtensionInterface
 {
-    public function load( array $configs, ContainerBuilder $container ): void
+    use PrependApiPlatformTrait;
+    
+    public function load( array $config, ContainerBuilder $container ): void
     {
-        $config = $this->processConfiguration( $this->getConfiguration( [], $container ), $configs );
+        $config = $this->processConfiguration( $this->getConfiguration( [], $container ), $config );
         
         $yamlLoader = new Loader\YamlFileLoader( $container, new FileLocator( __DIR__.'/../Resources/config' ) );
         $xmlLoader  = new Loader\XmlFileLoader( $container, new FileLocator( __DIR__ . '/../Resources/config' ) );
@@ -23,5 +26,15 @@ final class VSApiExtension extends Extension
         if ( $container->hasParameter( 'api_platform.enable_swagger_ui' ) && $container->getParameter( 'api_platform.enable_swagger_ui' ) ) {
             $xmlLoader->load( 'integrations/swagger.xml' );
         }
+        
+        $this->prepend( $container );
+    }
+    
+    public function prepend( ContainerBuilder $container ): void
+    {
+        $config = $container->getExtensionConfig( $this->getAlias() );
+        $config = $this->processConfiguration( $this->getConfiguration( [], $container ), $config );
+        
+        $this->prependApiPlatform( $container );
     }
 }